Move writer test programs to a specific folder
