+++ tests/mock/Programs/Writer/DeclareProcProgram.php
@@ -24,31 +24,38 @@
  * THE SOFTWARE.
  */
 
-namespace Guedel\Tests\Mock\AL\Programs;
+namespace Guedel\Tests\Mock\AL\Programs\Writer;
 
+use Guedel\Tests\Mock\AL\Programs\BaseTestProgram;
+
+use Guedel\AL\Declaration\ProcedureDecl;
+use Guedel\AL\Declaration\{Parameter, ParametersList};
 use Guedel\AL\Statement\ProcedureCall;
-use Guedel\AL\Statement\IfThenStmt;
-use Guedel\AL\Expression\Value;
+use Guedel\AL\Expression\Variable;
 
 /**
- * Description of SimpleIfProgram
+ * Description of DeclareProcProgram
  *
  * @author Guedel <guedel87@live.fr>
  */
-class SimpleIfProgram implements BaseTestProgram
+class DeclareProcProgram implements BaseTestProgram
 {
   //put your code here
   public function attend(): string
   {
     return join(PHP_EOL, [
-        "IF TRUE THEN",
-        "\tWRITE \"OK\"",
-        "END IF",
+        "PROCEDURE bonjour(IN name: ANY)",
+        "\tWRITE name",
+        "END PROCEDURE",
     ]) . PHP_EOL;
   }
 
   public function code(): \Guedel\AL\Statement\Statement
   {
-    return new IfThenStmt(new Value(true), new ProcedureCall("WRITE", new Value("OK")));
+    return new ProcedureDecl(
+        "bonjour",
+        new ParametersList(new Parameter("name", "IN")),
+        new ProcedureCall("WRITE", new Variable("name"))
+    );
   }
 }
