+++ tests/mock/Programs/Writer/FunctionDeclProgram.php
@@ -24,7 +24,9 @@
  * THE SOFTWARE.
  */
 
-namespace Guedel\Tests\Mock\AL\Programs;
+namespace Guedel\Tests\Mock\AL\Programs\Writer;
+
+use Guedel\Tests\Mock\AL\Programs\BaseTestProgram;
 
 use Guedel\AL\Statement\StatementList;
 use Guedel\AL\Declaration\FunctionDecl;
