+++ tests/mock/Programs/Writer/LoopProgram.php
@@ -24,36 +24,60 @@
  * THE SOFTWARE.
  */
 
-namespace Guedel\Tests\Mock\AL\Programs;
+namespace Guedel\Tests\Mock\AL\Programs\Writer;
 
-use Guedel\AL\Declaration\ProcedureDecl;
-use Guedel\AL\Declaration\{Parameter, ParametersList};
-use Guedel\AL\Statement\ProcedureCall;
-use Guedel\AL\Expression\Variable;
+use Guedel\Tests\Mock\AL\Programs\BaseTestProgram;
+
+use Guedel\AL\Declaration\VariableDecl;
+use Guedel\AL\Statement\{WhileStmt, ForStmt, ForEachStmt, StatementList, ProcedureCall};
+use Guedel\AL\Expression\{Value, Variable};
+use Guedel\AL\Datatype\ArrayOf;
+use Guedel\AL\Datatype\Any;
 
 /**
- * Description of DeclareProcProgram
+ * Description of LoopProgram
  *
  * @author Guedel <guedel87@live.fr>
  */
-class DeclareProcProgram implements BaseTestProgram
+class LoopProgram implements BaseTestProgram
 {
   //put your code here
   public function attend(): string
   {
     return join(PHP_EOL, [
-        "PROCEDURE bonjour(IN name: ANY)",
-        "\tWRITE name",
-        "END PROCEDURE",
+        "WHILE i DO",
+        "\tDEC i",
+        "END WHILE",
+        "FOR i FROM 1 TO 10 STEP 1 DO",
+        "\tWRITE i",
+        "END FOR",
+        "VAR arr: ARRAY OF ANY",
+        "FOR EACH item IN arr DO",
+        "\tWRITE item",
+        "END FOR",
     ]) . PHP_EOL;
   }
 
   public function code(): \Guedel\AL\Statement\Statement
   {
-    return new ProcedureDecl(
-        "bonjour",
-        new ParametersList(new Parameter("name", "IN")),
-        new ProcedureCall("WRITE", new Variable("name"))
+    return new StatementList(
+        new WhileStmt(
+            new Variable("i"),
+            new ProcedureCall("DEC", new Variable("i"))
+        ),
+        new ForStmt(
+            "i",
+            new Value(1),
+            new Value(10),
+            new Value(1),
+            new ProcedureCall("WRITE", new Variable("i"))
+        ),
+        new VariableDecl("arr", new ArrayOf(Any::getType())),
+        new ForEachStmt(
+            "item",
+            new Variable("arr"),
+            new ProcedureCall("WRITE", new Variable("item"))
+        )
     );
   }
 }
