+++ tests/mock/Programs/Writer/SimpleIfProgram.php
@@ -24,37 +24,33 @@
  * THE SOFTWARE.
  */
 
-namespace Guedel\Tests\Mock\AL\Programs;
+namespace Guedel\Tests\Mock\AL\Programs\Writer;
 
-use Guedel\AL\Statement\StatementList;
-use Guedel\AL\Statement\AssignStmt;
-use Guedel\AL\Declaration\VariableDecl;
+use Guedel\Tests\Mock\AL\Programs\BaseTestProgram;
+
+use Guedel\AL\Statement\ProcedureCall;
+use Guedel\AL\Statement\IfThenStmt;
 use Guedel\AL\Expression\Value;
-use Guedel\AL\Datatype\Any;
 
 /**
- * Description of VariableUseProgram
+ * Description of SimpleIfProgram
  *
  * @author Guedel <guedel87@live.fr>
  */
-class VariableUseProgram implements BaseTestProgram
+class SimpleIfProgram implements BaseTestProgram
 {
-  /**
-   * @inherit
-   */
+  //put your code here
   public function attend(): string
   {
     return join(PHP_EOL, [
-        "VAR i: ANY",
-        "i <- 10",
+        "IF TRUE THEN",
+        "\tWRITE \"OK\"",
+        "END IF",
     ]) . PHP_EOL;
   }
 
   public function code(): \Guedel\AL\Statement\Statement
   {
-    return new StatementList(
-        new VariableDecl("i"),
-        new AssignStmt("i", new Value(10))
-    );
+    return new IfThenStmt(new Value(true), new ProcedureCall("WRITE", new Value("OK")));
   }
 }
