+++ tests/mock/Programs/Writer/StructProgram.php
@@ -24,7 +24,9 @@
  * THE SOFTWARE.
  */
 
-namespace Guedel\Tests\Mock\AL\Programs;
+namespace Guedel\Tests\Mock\AL\Programs\Writer;
+
+use Guedel\Tests\Mock\AL\Programs\BaseTestProgram;
 
 use Guedel\AL\Datatype\{Structure, TypeName, Any};
 use Guedel\AL\Declaration\VariableDecl;
